Recursively checksum submodule content
Many projects use git submodules for "bundling" small libraries.  In
that case, because the submodule is really part of the source code, we
need to include the full contents in the checksum.

One could instead require that submodules have `Git-EVTag`, but part
of the idea of this use of submodules is to have a lightweight way to
pull in arbitrary versions of other code, without going through a more
heavyweight release process.

Where this recursion will be more painful is trying to use `Git-EVTag`
on an "agglomeration" project that pulls in many large codebases.
Using submodules instead of Android's `repo` tool for example.  But,
in that case, requiring `Git-EVTag` on each subproject makes more
sense to me - one would expect these projects to have real releases
with maintainers.