Refuse to tag/verify with dirty tree
For a future where we're handling submodules, we're going to require
that the checked out `HEAD` is the tag target.  This means that if the
submodule is modified, we're potentially processing the wrong data.

For creating tags, adding one to `HEAD` is the normal case, and in
the unusual case for tagging non-`HEAD`, having the user check out
first shouldn't be bad.

It's a bit more annoying for verifying tags; we're punting the problem
of getting checkouts to the caller.  But we can live with that.