Restrict tagging and verifying to HEAD
This is in preparation for submodule support.  The problem
with submodules is that in order to process the repository,
we need a checkout of the target (currently...it would likely
be possible to enhance libgit2 to not require this).

I'd rather not have git-evtag be in the business of setting up
temporary checkouts, so add this restriction now.

I expect it to not be too onerous in practice.  If it becomes a real
problem, we can always lift the restriction later with some work.