tests: Configure git author/committer name
Otherwise, recent versions of git will refuse to commit anything on an
unconfigured system.

Use Colin's name and email address here for now, rather than a dummy set
of data. This is because the EVTag covers these, so if we don't use the
same, we can't possibly get a match for the hard-coded expectation.

The expectation for the new test added by git-evtag#36 needs to be
adjusted for this, because it's comparing against a reference that
was generated using Philip Withnall's name and email address. It seems
simpler to use the same name and email address throughout the test suite.

Signed-off-by: Simon McVittie <smcv@debian.org>