tests: Configure git author/committer name
Otherwise, recent versions of git will refuse to commit anything on an
unconfigured system.

Use Colin's name and email address here for now, rather than a dummy set
of data. This is because the EVTag covers these, so if we don't use the
same, we can't possibly get a match for the hard-coded expectation.

Signed-off-by: Simon McVittie <smcv@debian.org>