Use g_spawn_check_wait_status instead of deprecated check_exit_status
The function we're calling always expected a wait-status parameter
as produced by g_spawn_sync() or waitpid(), not an exit status as
returned by WEXITSTATUS(), so its old name was misleading and was
deprecated in GLib 2.70. git-evtag correctly passes a wait status
to it.

In older GLib versions we can continue to use the deprecated name.

Signed-off-by: Simon McVittie <smcv@collabora.com>