tests: Include test-basic.sh in dist tarballs with newer Automake
The previous setup using $(addprefix) seems to be included in dist
tarballs by the Automake used in Gitlab-CI, but not by the Automake
in Debian unstable. Simplify the test setup so both work.

This also results in the .test files being generated in tests/ rather
than at top level, reducing clutter a bit.

Signed-off-by: Simon McVittie <smcv@debian.org>