tests: Fix installation directory for "as-installed" tests under Meson
The convention for "as-installed" tests is that they are in a
package-specific subdirectory of /usr/libexec/installed-tests, rather
than directly inside /usr/libexec/installed-tests.

The Autotools build used an installed-tests subdirectory of the
package-specific subdirectory of /usr/libexec, but this is
unconventional, so take the opportunity to behave more like other
packages.

Fixes: f6e1453f "build: Add a Meson build system"
Signed-off-by: Simon McVittie <smcv@debian.org>